fixed interfaces for playertable/alumni
